# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Graduation Project Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List my invitations
      responses:
        "200":
          description: Invitations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: sendInvitation
      tags: [invitations]
      summary: Send team invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInvitationBody"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /invitations/{id}/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /invitations/{id}/reject:
    post:
      operationId: rejectInvitation
      tags: [invitations]
      summary: Reject invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /phases/{teamId}:
    get:
      operationId: getTeamPhase
      tags: [phases]
      summary: Get team's current phase
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPhase"
  /phases/{teamId}/advance:
    post:
      operationId: advancePhase
      tags: [phases]
      summary: Advance team to next phase
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Phase advanced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPhase"
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meeting requests
      responses:
        "200":
          description: Meetings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: requestMeeting
      tags: [meetings]
      summary: Request a meeting with supervisor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestMeetingBody"
      responses:
        "201":
          description: Meeting requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}/approve:
    post:
      operationId: approveMeeting
      tags: [meetings]
      summary: Approve meeting request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}/reject:
    post:
      operationId: rejectMeeting
      tags: [meetings]
      summary: Reject meeting request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, supervisor, coordinator]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, supervisor, coordinator]
      required: [name, email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        projectTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [forming, active, supervised, completed]
        leaderId:
          type: integer
        supervisorId:
          type: integer
          nullable: true
        currentPhase:
          type: string
          enum: [proposal, progress, final]
          nullable: true
        createdAt:
          type: string
          format: date-time
        leader:
          $ref: "#/components/schemas/User"
        supervisor:
          $ref: "#/components/schemas/User"
          nullable: true
        memberCount:
          type: integer
      required: [id, name, status, leaderId, createdAt, leader, memberCount]
...
    Invitation:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        invitedUserId:
          type: integer
        invitedByUserId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
        team:
          $ref: "#/components/schemas/Team"
        invitedUser:
          $ref: "#/components/schemas/User"
        invitedBy:
          $ref: "#/components/schemas/User"
      required: [id, teamId, invitedUserId, invitedByUserId, status, createdAt, team, invitedUser, invitedBy]
    SendInvitationBody:
      type: object
      properties:
        teamId:
          type: integer
        invitedUserId:
          type: integer
      required: [teamId, invitedUserId]
...
    ProjectPhase:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        phase:
          type: string
          enum: [proposal, progress, final]
        status:
          type: string
          enum: [in_progress, completed]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        team:
          $ref: "#/components/schemas/Team"
      required: [id, teamId, phase, status, startedAt, team]
...
    Meeting:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        requestedById:
          type: integer
        supervisorId:
          type: integer
        proposedDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        team:
          $ref: "#/components/schemas/Team"
        requestedBy:
          $ref: "#/components/schemas/User"
        supervisor:
          $ref: "#/components/schemas/User"
      required: [id, teamId, requestedById, supervisorId, proposedDate, status, createdAt, team, requestedBy, supervisor]
    RequestMeetingBody:
      type: object
      properties:
        supervisorId:
          type: integer
        proposedDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required: [supervisorId, proposedDate]
...